requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
celery==5.4.0
redis==5.0.7
python-dotenv==1.0.1
